Market work

+++ core/src/com/mygdx/game/objects/Transaction.kt
@@ -1,7 +1,7 @@
 package com.mygdx.game.objects
 
-class Transaction(val name:String, val amount:Int, val pricePerUnit:Int) {
-    val totalPrice:Int
+class Transaction(val name:String, val amount:Int, val pricePerUnit:Float) {
+    val totalPrice:Float
         get() = amount*pricePerUnit
 
     override fun toString(): String {
